Random: Prevent URL injection in 'picsum' command
While we can URL-encode the input string and try to make it safe, it's
much easier to just hash it. It's literally just a seed and it's all
randomized by message's ID anyways.

To be honest, it is only to make the users feel like they have some
controll over their lives and actions they are taking.

:)

Resolves #60